feat: Add new settings route
